+++ ElvUI_LuckyoneUI/Core/Layout.lua
@@ -351,6 +351,7 @@ if L1UI.Retail then
 		E.db["chat"]["showHistory"]["SAY"] = false
 		E.db["chat"]["showHistory"]["YELL"] = false
 		E.db["chat"]["tabFont"] = "Expressway"
+		E.db["chat"]["tabFontOutline"] = "OUTLINE"
 		E.db["chat"]["tabFontSize"] = 11
 		E.db["chat"]["tabSelectedTextEnabled"] = false
 		E.db["chat"]["tabSelector"] = "NONE"
@@ -2227,6 +2228,7 @@ elseif L1UI.Classic then
 		E.db["chat"]["panelWidthRight"] = 414
 		E.db["chat"]["separateSizes"] = true
 		E.db["chat"]["tabFont"] = "Expressway"
+		E.db["chat"]["tabFontOutline"] = "OUTLINE"
 		E.db["chat"]["tabFontSize"] = 11
 		E.db["chat"]["tabSelectedTextEnabled"] = false
 		E.db["chat"]["tabSelector"] = "NONE"
@@ -3520,6 +3522,7 @@ elseif L1UI.TBC then
 		E.db["chat"]["panelWidthRight"] = 414
 		E.db["chat"]["separateSizes"] = true
 		E.db["chat"]["tabFont"] = "Expressway"
+		E.db["chat"]["tabFontOutline"] = "OUTLINE"
 		E.db["chat"]["tabFontSize"] = 11
 		E.db["chat"]["tabSelectedTextEnabled"] = false
 		E.db["chat"]["tabSelector"] = "NONE"